+++ src/seedcase_sprout/core/write_resource_properties.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from seedcase_sprout.core.check_is_file import check_is_file
 from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
-from seedcase_sprout.core.deep_update import nested_update
+from seedcase_sprout.core.nested_update import nested_update
 from seedcase_sprout.core.properties import ResourceProperties
 from seedcase_sprout.core.read_json import read_json
 from seedcase_sprout.core.sprout_checks.check_properties import check_properties
